rearranged some code
