+++ contracts/distribution/dao-rewards-distributor/src/lib.rs
@@ -4,6 +4,7 @@ pub mod contract;
 mod error;
 pub mod hooks;
 pub mod msg;
+pub mod rewards;
 pub mod state;
 
 #[cfg(test)]
